requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
APScheduler==3.10.4